requirements.txt
azure-cognitiveservices-speech==1.9.0
language-check==1.1
pydub==0.23.1
pyaudio==0.2.11
pymongo==3.10.1
dnspython==1.16.0
textblob==0.15.3
pillow==7.0.0